[#190]refactor: 세션찾기 오류 및 아고라 멤버찾기 오류 해결

+++ src/main/java/com/attica/athens/domain/member/application/MemberService.java
@@ -4,10 +4,13 @@ import com.attica.athens.domain.member.dao.MemberRepository;
 import com.attica.athens.domain.member.domain.Member;
 import com.attica.athens.domain.member.dto.request.CreateMemberRequest;
 import com.attica.athens.domain.member.dto.response.GetMemberResponse;
+import com.attica.athens.domain.member.exception.AlreadyActivateMemberException;
 import com.attica.athens.domain.member.exception.DuplicateMemberException;
 import com.attica.athens.domain.member.exception.NotFoundMemberException;
+import com.attica.athens.domain.member.exception.UnauthorizedException;
 import com.attica.athens.global.auth.application.AuthService;
 import com.attica.athens.global.auth.domain.AuthProvider;
+import com.attica.athens.global.auth.domain.CustomUserDetails;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -55,9 +58,13 @@ public class MemberService {
     }
 
     @Transactional
-    public void deleteMember(Long memberId) {
+    public void deleteMember(Long memberId, CustomUserDetails userDetails) {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new NotFoundMemberException(memberId));
+
+        if (!memberId.equals(userDetails.getUserId())) {
+            throw new UnauthorizedException();
+        }
 
         member.delete();
     }
@@ -66,6 +73,10 @@ public class MemberService {
     public void restoreMember(Long memberId) {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new NotFoundMemberException(memberId));
+
+        if (!member.isDeleted()) {
+            throw new AlreadyActivateMemberException();
+        }
 
         member.restore();
     }
